fix: fix function call gas assignment
When deploying big (1MB) contract the create call failed, even though
the supplied amount was 300 TGas. By experimenting it was discovered
that function_call done right after the deploy subtracts Gas available
to the current method. However this gas limit is NOT used when the
actual init method is called. The solution is to set function call gas
to a intentionally low value (1 TGas) to stop it from failing.